# TraceQL performance improvements