--- a/pkg/traceql/enum_statics.go
+++ b/pkg/traceql/enum_statics.go
@@ -174,6 +174,7 @@ func (k Kind) String() string {
 }
 
 var (
+	StaticNil   = NewStaticNil()
 	StaticTrue  = NewStaticBool(true)
 	StaticFalse = NewStaticBool(false)
 )