@mdisibio 2024-09-26
For completeness ran benchmarks on metrics.   Nice improvement to the compare() function, some smaller wins for other queriers, although harder to see because they are noiser:
```
name                                                                       old time/op    new time/op    delta
BackendBlockQueryRange/{}_|_rate()/7                                          2.51s ± 9%     2.34s ± 4%     ~     (p=0.093 n=6+6)
BackendBlockQueryRange/{}_|_rate()_by_(name)/7                                4.33s ± 3%     4.28s ± 6%     ~     (p=0.937 n=6+6)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7               3.30s ± 7%     3.17s ± 5%     ~     (p=0.132 n=6+6)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7                       5.99s ± 8%     5.65s ± 3%   -5.69%  (p=0.026 n=6+6)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7     323ms ± 3%     319ms ± 2%     ~     (p=0.662 n=6+5)
BackendBlockQueryRange/{status=error}_|_rate()/7                              286ms ± 1%     284ms ± 2%     ~     (p=0.240 n=6+6)
BackendBlockQueryRange/{_}_|_compare({status=error})/7                        82.4s ± 6%     77.7s ±19%     ~     (p=0.126 n=5+6)

name                                                                       old MB_IO/op   new MB_IO/op   delta
BackendBlockQueryRange/{}_|_rate()/7                                           30.5 ± 0%      30.5 ± 0%     ~     (all equal)
BackendBlockQueryRange/{}_|_rate()_by_(name)/7                                 37.4 ± 0%      37.4 ± 0%     ~     (all equal)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7                31.0 ± 0%      31.0 ± 0%     ~     (all equal)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7                        33.5 ± 0%      33.5 ± 0%     ~     (all equal)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7      31.0 ± 0%      31.0 ± 0%     ~     (all equal)
BackendBlockQueryRange/{status=error}_|_rate()/7                               32.0 ± 0%      32.0 ± 0%     ~     (all equal)
BackendBlockQueryRange/{_}_|_compare({status=error})/7                          220 ± 0%       220 ± 0%     ~     (all equal)

name                                                                       old spans/op   new spans/op   delta
BackendBlockQueryRange/{}_|_rate()/7                                          5.40M ± 0%     5.40M ± 0%     ~     (all equal)
BackendBlockQueryRange/{}_|_rate()_by_(name)/7                                5.40M ± 0%     5.40M ± 0%     ~     (all equal)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7               5.40M ± 0%     5.40M ± 0%     ~     (all equal)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7                       5.40M ± 0%     5.40M ± 0%     ~     (all equal)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7      111k ± 0%      111k ± 0%     ~     (all equal)
BackendBlockQueryRange/{status=error}_|_rate()/7                              34.1k ± 0%     34.1k ± 0%     ~     (all equal)
BackendBlockQueryRange/{_}_|_compare({status=error})/7                        5.40M ± 0%     5.40M ± 0%     ~     (all equal)

name                                                                       old spans/s    new spans/s    delta
BackendBlockQueryRange/{}_|_rate()/7                                          2.16M ± 9%     2.31M ± 4%     ~     (p=0.093 n=6+6)
BackendBlockQueryRange/{}_|_rate()_by_(name)/7                                1.25M ± 3%     1.26M ± 6%     ~     (p=0.937 n=6+6)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7               1.64M ± 7%     1.71M ± 6%     ~     (p=0.132 n=6+6)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7                        903k ± 7%      956k ± 3%   +5.86%  (p=0.026 n=6+6)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7      343k ± 3%      347k ± 2%     ~     (p=0.662 n=6+5)
BackendBlockQueryRange/{status=error}_|_rate()/7                               119k ± 1%      120k ± 2%     ~     (p=0.240 n=6+6)
BackendBlockQueryRange/{_}_|_compare({status=error})/7                        65.6k ± 6%     70.0k ±16%     ~     (p=0.126 n=5+6)

name                                                                       old alloc/op   new alloc/op   delta
BackendBlockQueryRange/{}_|_rate()/7                                         11.5MB ± 8%    11.5MB ± 8%     ~     (p=1.000 n=5+5)
BackendBlockQueryRange/{}_|_rate()_by_(name)/7                               16.7MB ±36%    15.5MB ± 9%     ~     (p=1.000 n=5+5)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7              12.0MB ± 7%    12.0MB ± 7%     ~     (p=0.952 n=5+5)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7                       389MB ± 4%     385MB ± 3%     ~     (p=0.421 n=5+5)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7    46.6MB ± 1%    46.7MB ± 1%     ~     (p=0.699 n=6+6)
BackendBlockQueryRange/{status=error}_|_rate()/7                             3.51MB ± 0%    4.19MB ±30%     ~     (p=0.329 n=5+6)
BackendBlockQueryRange/{_}_|_compare({status=error})/7                       8.19GB ± 5%    4.13GB ± 2%  -49.57%  (p=0.008 n=5+5)

name                                                                       old allocs/op  new allocs/op  delta
BackendBlockQueryRange/{}_|_rate()/7                                          46.8k ± 0%     46.8k ± 0%     ~     (p=1.000 n=5+5)
BackendBlockQueryRange/{}_|_rate()_by_(name)/7                                65.3k ±38%     59.1k ± 0%     ~     (p=0.937 n=5+5)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7               48.8k ± 0%     48.8k ± 0%     ~     (p=0.595 n=5+5)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7                       1.01M ± 7%     0.98M ± 1%     ~     (p=0.126 n=6+5)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7      250k ± 0%      250k ± 0%     ~     (p=0.329 n=6+5)
BackendBlockQueryRange/{status=error}_|_rate()/7                              61.0k ± 0%     61.1k ± 0%     ~     (p=0.104 n=5+6)
BackendBlockQueryRange/{_}_|_compare({status=error})/7                        66.2M ± 7%     32.2M ± 0%  -51.35%  (p=0.008 n=5+5)
```